2025 Idea addition: Extend BMv2 to output all possible outcomes
New for this year, but would be wildly useful to us at Google at least. I don't imagine it would be too difficult, except perhaps extending the output type (and thus perhaps input types in various parts of the pipeline). Perhaps some of that can be hacked away by providing a new function that gets/sends packets in this way instead of changing the type of current functions.

Signed-off-by: Jonathan DiLorenzo <jonathan.dilorenzo@gmail.com>